+++ CHANGELOG.md
@@ -28,7 +28,7 @@ CHANGELOG
 - `intelmq.bots.collectors.eset.collector`: Added (PR#1554 by Mikk Margus Möll).
 - `intelmq.bots.collectors.http.collector_http`:
   - Added PGP signature check functionality (PR#1602 by sinus-x).
-  - If status code is not 2xx, the response header and body are logged in debug logging level (#1615).
+  - If status code is not 2xx, the request's and response's headers and body are logged in debug logging level (#1615).
 
 #### Parsers
 - `intelmq.bots.parsers.eset.parser`: Added (PR#1554 by Mikk Margus Möll).